Version 1.
Adding to github from local repo.

+++ src/main/java/com/example/restservice/controller/ContentController.java
@@ -8,13 +8,10 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-
-
-@RestController // Handel request mapping
+@RestController
 @RequestMapping("/api/content")
 @CrossOrigin
 public class ContentController {
-//Accepting req and return response
 
     private final ContentCollectionRepo repo;
 
@@ -22,28 +19,24 @@ public class ContentController {
         this.repo = repo;
     }
 
-    @GetMapping("") //Lists all content
+    @GetMapping("")
     public List<Content>findAll() {
+
         return repo.findAll();
     }
 
-    //Create, Read, Update, Delete- C.R.U.D || filter-sorting
-
-    //Method for getting by certain id
     @GetMapping("/{id}")
     public Content findById(@PathVariable Integer id){
         return repo.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Content not found"));
     }
 
-    @ResponseStatus(HttpStatus.CREATED) //Creates 201...
+    @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("")
     public void create(@Valid @RequestBody Content content){
         repo.save(content);
     }
 
-
-    //http://localhost:8080/api/content/2
-    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @ResponseStatus(HttpStatus.OK)
     @PutMapping("/{id}")
     public void update(@Valid @RequestBody Content content, @PathVariable Integer id){
 
@@ -56,6 +49,10 @@ public class ContentController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Integer id){
+
+        if (!repo.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Content not found!");
+        }
         repo.delete(id);
     }
 }
